+++ packages/isaacscript-common/src/functions/array.ts
@@ -361,18 +361,6 @@ export function getHighestArrayElement(array: number[]): number | undefined {
 }
 
 /**
- * Helper function to return the last element of an array.
- *
- * If the array is empty, this will return undefined.
- *
- * (Note that TSTL does not support `Array.at(-1)`, which would make this helper function largely
- * unnecessary.)
- */
-export function getLastElement<T>(array: T[]): T | undefined {
-  return array[array.length - 1];
-}
-
-/**
  * Helper function to get the lowest value in an array. Returns undefined if there were no elements
  * in the array.
  */